requirements.txt
accelerate==0.24.1
cudatoolkit==11.7
datasets==2.14.6
debugpy==1.8.0
deepspeed==0.13.2
gensim==4.3.2
ipykernel==6.26.0
loompy==3.0.7
matplotlib==3.7.3
multiprocess==0.70.15
networkx==3.1
numpy==1.24.3
pandas==2.0.3
python==3.8.18
pytorch==2.0.1
ray==2.9.0
scanpy==1.9.6
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
statsmodels==0.14.0
tensorboard==2.14.0
tokenizers==0.15.0
torchaudio==2.0.2
torchmetrics==1.2.1
torchvision==1.2.1
transformers==4.36.2
